update package.sh to remove pre-built db for starter project only.

+++ modules/ROOT/examples/package.sh
@@ -5,15 +5,25 @@ if [[ ! -z ${1} ]]; then
 		echo "Temporary folder..."
 		mkdir $OUTPUT_FILE
 		cp -R HotelFinder "${OUTPUT_FILE}/"
-		echo "Clean..."
+		echo "Clean... ${1}"
+
+		if [ "${1}" = "starter-project" ]; then
+			rm -rf "${OUTPUT_FILE}/HotelFinder/ios/HotelFinder/travel-sample.cblite2"
+			rm -rf "${OUTPUT_FILE}/HotelFinder/android/app/src/main/assets/travel-sample.cblite2.zip"
+		elif [ "${1}" = "final-project" ]; then
+			echo ""
+		else
+			echo "Error: the output file must be called starter-project or final-project"
+		fi
+
 		rm -rf "${OUTPUT_FILE}/HotelFinder/node_modules/"
 		rm -rf "${OUTPUT_FILE}/HotelFinder/ios/Frameworks/CouchbaseLiteSwift.framework"
-		rm -rf "${OUTPUT_FILE}/HotelFinder/android/app/src/main/assets/travel-sample.cblite2.zip"
-		rm -rf "${OUTPUT_FILE}/HotelFinder/ios/HotelFinder/travel-sample.cblite2.zip"
+		rm -rf "${OUTPUT_FILE}/HotelFinder/android/app/build"
 		rm -rf "${OUTPUT_FILE}/HotelFinder/android/.idea"
 		rm -rf "${OUTPUT_FILE}/HotelFinder/android/.gradle"
 		rm -rf "${OUTPUT_FILE}/HotelFinder/android/local.properties"
 		rm -rf "${OUTPUT_FILE}/HotelFinder/android/HotelFinder.iml"
+
 		echo "Zipping..."
 		ditto -c -k --sequesterRsrc --keepParent "${OUTPUT_FILE}" "${OUTPUT_FILE}.zip"
 		echo "Copy..."
